+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+**Changed**
+
+- Refactor `apics::gateway_props` defined type
+
 ## Release 0.5.3
 
 **Changed**